Move entrance selector code to Grid class
The code that selects entrances should live on the grid class so I've
moved it there.

There is more logic still that should move such as the
actual code that draws the lines onto the canvas that currently lives
inside each generator algorithms `draw()` method.

In the future the `draw()` method will only be for actually drawing the
maze and that code will live on the Grid class. The stuff that currently
serves to "generate" the maze will be pulled out into a new method
called generate()/build()/execute().