Fixed corner cell entrance issue
The previous code was simply checking our entrances map to see if it
contained the ID of the current cell and then skipping that cell
altogether if it was indeed an entrance.

This caused an issue when the cell lived in one of the four corners of
our maze/grid by way of not having either it's West/North, North/East,
West/South, or East/South borders drawn. This appeared visually odd
because the entire corner would be missing. This didn't affect the
solvability of the maze at all.

Now we store exactly which border should be skipped in our entrances map
object and then use that to only skip drawing the border line for the
direction that was chosen when we selected the entrances.