Implement `create()` for Sidewinder
Since I'm now using `create()` as the actual method that runs the "meat
and potatoes" of the algorithm and then reserving `draw()` on the Grid
class to actually render the maze to the screen, here's this commit.

I'm not sure if I previously had some kind of caching issue where the
change I made to `shouldCloseRun` is now a 50% chance to close the run
instead of the previous 33.33% chance but in making this change/refactor
the Sidewinder algorithm now seems to be generating correct mazes.

Cheers to that.