Show cell weight scores based on config
If you pass a config object to the contructor for Dijkstra with a
property called `showWeightedScores: true` it will make the grid render
the weight of the cell when assigning distances from the starting point.

A cells distance score or weight is just the number of cells it is away
from the starting point when moving in one of the cardinal directions
(no diagonals).