`fillRect()` on the end cell
Previously we weren't painting/coloring the ending cell of our solution,
that is now fixed by checking if the cell passed to `fillShortestPath()`
is the end cell.

Also fixed an issue with x,y being set inside the loop on cell links to
the row/col of the cell instead of the current link.